requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.13.0,<0.24.0
python-dotenv>=0.19.0,<2.0.0
aiohttp>=3.8.0,<4.0.0
google-auth>=2.3.0,<3.0.0
google-cloud-aiplatform>=1.24.0,<1.25.0  # 使用兼容版本
protobuf>=3.0.0,<4.0.0
colorlog>=6.7.0,<7.0.0
python-multipart>=0.0.5,<0.0.6  # 用于处理表单数据
pydantic>=1.8.0,<2.0.0  # FastAPI的依赖项 ni